Add Multiples!
This update adds what I'm calling multiples.
A multiple is basically the combination of recipes.
For ex: In the crucible you have a recipe for lava from 12 of EACH terracotta type. Previously that was a recipe for each. Now it will be a single recipe that cycles through them and when you hover shows you all of the possible inputs.
Also bug fixes :thumbsup: